requirements.txt
streamlit
google-cloud-storage
google-cloud-bigquery
db-dtypes
pandas
selenium>=4.10.0